Этап 6: финальная проверка перед сдачей
Статическая проверка:
- ruff подключён с явным набором правил (E,F,I,UP,B,C4,SIM,RUF)
- исправлены все 69 замечаний: порядок импортов, устаревшие
  конструкции, изменяемые атрибуты классов, лишние приведения типов,
  подавление исключений через contextlib, itertools.pairwise
- ruff добавлен в requirements

Найдено и исправлено:
- тесты интерфейса писали в общий файл конфигурации симулятора
  в каталоге временных файлов пользователя и зависели от его
  содержимого: сохранённое одним прогоном значение ломало следующий.
  Хранилище изолируется фикстурой на каждый тест
- лямбда в цикле обращалась к переменной, удаляемой в конце итерации
- demo_cli выводил кириллицу в кодировке консоли Windows

CLAUDE.md исключён из поставки: рабочие заметки, не предназначенные
для передачи вместе с проектом.

Проверено: 345 быстрых тестов (10 прогонов подряд), 15 нагрузочных,
26 измерений без отклонений, сборка прошивки, сборка и самопроверка
исполняемого файла, сценарий в консоли.

+++ desktop_app/servo_configurator/simulation/servo_model.py
@@ -178,15 +178,15 @@ class ServoModel:
         assert rng is not None
 
         load = self.load + rng.gauss(0.0, _LOAD_NOISE)
-        signed_load = int(round(load)) * (1 if self.velocity >= 0 else -1)
+        signed_load = round(load) * (1 if self.velocity >= 0 else -1)
 
         return {
-            "pos": int(round(self.position)) + rng.choice((-1, 0, 0, 0, 1)),
-            "spd": int(round(self.velocity)),
+            "pos": round(self.position) + rng.choice((-1, 0, 0, 0, 1)),
+            "spd": round(self.velocity),
             "load": max(-1000, min(1000, signed_load)),
             "volt": self.voltage_dv + rng.choice((-1, 0, 0, 1)),
             "temp": self.temperature_c,
-            "cur": int(round(abs(load) * _CURRENT_PER_LOAD)),
+            "cur": round(abs(load) * _CURRENT_PER_LOAD),
             "moving": abs(self.velocity) > _MOVING_THRESHOLD,
         }
